fixed bugs

+++ db/migrate/20200515125828_create_warehouses.rb
@@ -2,7 +2,7 @@ class CreateWarehouses < ActiveRecord::Migration[6.0]
   def change
     create_table :warehouses do |t|
       t.references :product, null: false, foreign_key: true
-      t.integer :amount
+      t.decimal :amount
 
       t.timestamps
     end
